with backup stuff

+++ server/server.js
@@ -92,62 +92,69 @@ app.post('/register', async (req,res) => {
   var client_ip = decrypt(Buffer.from(req.body.ip), sessionKey);
   var client_public_key = decrypt(Buffer.from(req.body.publickey), sessionKey);
 
-  if(isValidIP(client_ip)){
+  // Check for concurrency errors with the relay session key
+  if (client_type == -1 || client_pin == -1 || target_pin == -1 || client_ip == -1  || client_public_key == -1) { 
+    console.log("[!][Establish] RSK Concurrency Error"); 
+    res.send("RSK Concurrency Error"); 
+  }
+  else {
+    if(isValidIP(client_ip)){
 
-    // Check for duplicates
-    var sql = "SELECT pin FROM sessions WHERE pin=" +client_pin+";";
-    sqlConnection.query(sql, function(err, result) {
-      if(err) throw err;
-
-      //If duplicate PIN, delete the existing entry - because publickey/ip might have changed
-      if (result.length > 0 ) {
-        console.log("Duplicate PIN, updating entry.");
-        plainSQL("DELETE FROM sessions WHERE pin="+client_pin+";");
-      }
-
-      // Store this session entry
-      var sql = "INSERT INTO sessions (pin, targetPIN, publickey, ip, usertype) " 
-          +"VALUES (" + client_pin + ","+ target_pin + ",'" + client_public_key + "','" + client_ip + "','" + client_type + "')";
-      sqlConnection.query(sql, function (err, result) {   
-        if (err) throw err;
-        console.log('[+] Added session:\n      PIN:', client_pin, '\n     tPIN:', target_pin, '\n     Type:',client_type,
-                    '\n      PBK:',client_public_key,'\n       IP:',client_ip, '\n');
-
-        var match_pin, match_ip, match_pbk;
-
-        checkForMatch(client_pin,target_pin,client_type).then((match) =>{
-          if(match!=null){
-            match_pin = match[0];
-            match_ip = match[1];
-            match_pbk = match[2];
-          
-            var encrypted_match_pin = encrypt(match_pin.toString(),sessionKey);
-            var encrypted_match_ip = encrypt(match_ip.toString(),sessionKey);
-            var encrypted_match_pbk = encrypt(match_pbk.toString(),sessionKey);
-
-            console.log("[->] Sending:\n      PIN: "+encrypted_match_pin.toString('hex')+
-                  "\n       IP: "+encrypted_match_ip.toString('hex')+"\n      PBK: "+encrypted_match_pbk.toString('hex')+'\n');
-
-            plainSQL("DELETE FROM sessions WHERE pin="+target_pin+";");
-            res.json({ pin: encrypted_match_pin, ip: encrypted_match_ip, pbk: encrypted_match_pbk });
-
-            // CANCELLED?: Exchange the info to the peers so they can establish a sessionKey - they store it 'permanently' in datastores
-            // OK for this peer he's connected.. what about the other peer? i have his IP sure, but how do i actually CONNECT
-            // maybe enforce that he has to be connected (TCP) but then ok how do i find him while he's connected [session/cookies?]
-          }
-        }).catch(error => {
-          console.log(error);
-          if(error=="No match found"){
-            res.send("AWAITMATCH");
-          }
+      // Check for duplicates
+      var sql = "SELECT pin FROM sessions WHERE pin=" +client_pin+";";
+      sqlConnection.query(sql, function(err, result) {
+        if(err) throw err;
+  
+        //If duplicate PIN, delete the existing entry - because publickey/ip might have changed
+        if (result.length > 0 ) {
+          console.log("Duplicate PIN, updating entry.");
+          plainSQL("DELETE FROM sessions WHERE pin="+client_pin+";");
+        }
+  
+        // Store this session entry
+        var sql = "INSERT INTO sessions (pin, targetPIN, publickey, ip, usertype) " 
+            +"VALUES (" + client_pin + ","+ target_pin + ",'" + client_public_key + "','" + client_ip + "','" + client_type + "')";
+        sqlConnection.query(sql, function (err, result) {   
+          if (err) throw err;
+          console.log('[+] Added session:\n      PIN:', client_pin, '\n     tPIN:', target_pin, '\n     Type:',client_type,
+                      '\n      PBK:',client_public_key,'\n       IP:',client_ip, '\n');
+  
+          var match_pin, match_ip, match_pbk;
+  
+          checkForMatch(client_pin,target_pin,client_type).then((match) =>{
+            if(match!=null){
+              match_pin = match[0];
+              match_ip = match[1];
+              match_pbk = match[2];
+            
+              var encrypted_match_pin = encrypt(match_pin.toString(),sessionKey);
+              var encrypted_match_ip = encrypt(match_ip.toString(),sessionKey);
+              var encrypted_match_pbk = encrypt(match_pbk.toString(),sessionKey);
+  
+              console.log("[->] Sending:\n      PIN: "+encrypted_match_pin.toString('hex')+
+                    "\n       IP: "+encrypted_match_ip.toString('hex')+"\n      PBK: "+encrypted_match_pbk.toString('hex')+'\n');
+  
+              plainSQL("DELETE FROM sessions WHERE pin="+target_pin+";");
+              res.json({ pin: encrypted_match_pin, ip: encrypted_match_ip, pbk: encrypted_match_pbk });
+  
+              // CANCELLED?: Exchange the info to the peers so they can establish a sessionKey - they store it 'permanently' in datastores
+              // OK for this peer he's connected.. what about the other peer? i have his IP sure, but how do i actually CONNECT
+              // maybe enforce that he has to be connected (TCP) but then ok how do i find him while he's connected [session/cookies?]
+            }
+          }).catch(error => {
+            console.log(error);
+            if(error=="No match found"){
+              res.send("AWAITMATCH");
+            }
+          });
+  
         });
-
+        
       });
-      
-    });
-  } else {
-    console.log("Invalid IP");
-    res.send('ERROR');
+    } else {
+      console.log("Invalid IP");
+      res.send('ERROR');
+    }
   }
 });
 
@@ -156,38 +163,49 @@ app.post('/awaitMatch', async (req,res) => {
   var client_pin = decrypt(Buffer.from(req.body.pin), sessionKey);
   var target_pin = decrypt(Buffer.from(req.body.targetpin), sessionKey);
 
-  await checkForMatch(client_pin,target_pin,client_type).then((match) => {
+  if (client_type == -1 || client_pin == -1 || target_pin == -1) {
+    console.log("[!][awaitMatch] RSK Concurrency Error"); 
+    res.send("RSK Concurrency Error");
+  }
+  else {
+    await checkForMatch(client_pin,target_pin,client_type).then((match) => {
 
-    if(match.length==3){
-      match_pin = match[0];
-      match_ip = match[1];
-      match_pbk = match[2];
-    
-      var encrypted_match_pin = encrypt(match_pin.toString(),sessionKey);
-      var encrypted_match_ip = encrypt(match_ip.toString(),sessionKey);
-      var encrypted_match_pbk = encrypt(match_pbk.toString(),sessionKey);
+      if(match.length==3){
+        match_pin = match[0];
+        match_ip = match[1];
+        match_pbk = match[2];
+      
+        var encrypted_match_pin = encrypt(match_pin.toString(),sessionKey);
+        var encrypted_match_ip = encrypt(match_ip.toString(),sessionKey);
+        var encrypted_match_pbk = encrypt(match_pbk.toString(),sessionKey);
 
-      console.log("[->] Sending:\n      PIN: "+encrypted_match_pin.toString('hex')+
-            "\n       IP: "+encrypted_match_ip.toString('hex')+"\n      PBK: "+encrypted_match_pbk.toString('hex')+'\n');
+        console.log("[->] Sending:\n      PIN: "+encrypted_match_pin.toString('hex')+
+              "\n       IP: "+encrypted_match_ip.toString('hex')+"\n      PBK: "+encrypted_match_pbk.toString('hex')+'\n');
 
-      plainSQL("DELETE FROM sessions WHERE pin="+target_pin+";");
-      res.json({ pin: encrypted_match_pin, ip: encrypted_match_ip, pbk: encrypted_match_pbk });
+        plainSQL("DELETE FROM sessions WHERE pin="+target_pin+";");
+        res.json({ pin: encrypted_match_pin, ip: encrypted_match_ip, pbk: encrypted_match_pbk });
 
-    }
-    else res.send("NOMATCH"); //this might cause issues with databoxui
+      }
+      else res.send("NOMATCH"); //this might cause issues with databoxui
 
-  }).catch(error => {
-      console.log(error);
-      res.send("ERROR");
-  });
-    
+    }).catch(error => {
+        console.log(error);
+        res.send("ERROR");
+    });
+  }
 });
 
 app.post('/deleteSessionInfo', (req,res) => {
   var client_pin = decrypt(Buffer.from(req.body.pin), sessionKey);
-  plainSQL("DELETE FROM sessions WHERE pin="+client_pin+";");
-  console.log("Deleted session entry");
-  res.end();
+  if (client_pin == -1) { 
+    console.log("[!][Delete] RSK Concurrency Error"); 
+    res.send("RSK Concurrency Error");
+  }
+  else {
+    plainSQL("DELETE FROM sessions WHERE pin="+client_pin+";");
+    console.log("Deleted session entry");
+    res.end();
+  }
 });
 
 /****************************************************************************
@@ -195,74 +213,68 @@ app.post('/deleteSessionInfo', (req,res) => {
 ****************************************************************************/
 app.post('/retrieve', (req,res) =>{
   var pin = decrypt(Buffer.from(req.body.pin), sessionKey);
-  sqlConnection.query("SELECT data, checksum FROM databoxrhm WHERE pin=?;", [pin], function (err, rows) {
-    if(rows!=null && rows!=[] && rows.length>0){
-      var result = [];
-      for (var i = 0;i < rows.length; i++) {
-        var data = rows[i].data;
-        var checksum = rows[i].checksum;
-        var entry = {checksum,data};
-        result.push(entry);
+  if (pin==-1) { 
+    console.log("[!][Retrieve] RSK Concurrency Error"); 
+    res.send("RSK Concurrency Error");
+  }
+  else {
+    sqlConnection.query("SELECT data, checksum FROM databoxrhm WHERE pin=?;", [pin], function (err, rows) {
+      if(rows!=null && rows!=[] && rows.length>0){
+        var result = [];
+        for (var i = 0;i < rows.length; i++) {
+          var data = rows[i].data;
+          var checksum = rows[i].checksum;
+          var entry = {checksum,data};
+          result.push(entry);
+        }
+        if(result.length == 0) res.send('No data found.'); // send EOF empty array
+        
+        console.log("Found patient data:",result);
+        res.send(result);
+        //more checking before doing this tho... confirm that they got it and decrypted it
+        sqlConnection.query("DELETE FROM databoxrhm WHERE pin=?;",[pin]);
       }
-      if(result.length == 0) res.send('No data found.'); // send EOF empty array
-      
-      console.log("Found patient data:",result);
-      res.send(result);
-      //more checking before doing this tho... confirm that they got it and decrypted it
-      sqlConnection.query("DELETE FROM databoxrhm WHERE pin=?;",[pin]);
-    }
-    else {
-      var empty = [];
-      empty.push('EOF');
-      res.send(empty);
-    }
-  });
+      else {
+        var empty = [];
+        empty.push('EOF');
+        res.send(empty);
+      }
+    });
+  }
 });
 
 app.post('/store', (req,res) =>{
   var pin = decrypt(Buffer.from(req.body.pin),sessionKey);
   //var ttl = decrypt('aes-256-cbc', sessionKey, Buffer.from(req.body.ttl));
-  var data = Buffer.from(req.body.data); //dont try to decrypt - crashes cause doesnt have peerkey
-  var checksum = Buffer.from(req.body.checksum);
-
-  sqlConnection.query("INSERT INTO databoxrhm (pin, checksum, data) VALUES (?, ?,?);", [pin,checksum,data], function (err, result) { 
-    if (result!=null) console.log('[+] Added data:\n      PIN:', pin,'\n checksum:', checksum,'\n     data:', data);
-  });
-
-  res.send("ok");
+  if (pin==-1) { 
+    console.log("[!][Store] RSK Concurrency Error"); 
+    res.send("RSK Concurrency Error");
+  }
+  else {
+    var data = Buffer.from(req.body.data); //dont try to decrypt - crashes cause doesnt have peerkey
+    var checksum = Buffer.from(req.body.checksum);
+  
+    sqlConnection.query("INSERT INTO databoxrhm (pin, checksum, data) VALUES (?, ?,?);", [pin,checksum,data], function (err, result) { 
+      if (result!=null) console.log('[+] Added data:\n      PIN:', pin,'\n checksum:', checksum,'\n     data:', data);
+    });
+  
+    res.send("ok");
+  }
 });
 
 /****************************************************************************
 * Encrypt / Decrypt
 ****************************************************************************/
-
-
-
-// USE TRY CATCH because might have the concurrent request issue and bad_decrypt
-// on error, send back to client to re-attempt and client auto re-sends
-
-
-
 //based on https://lollyrock.com/posts/nodejs-encryption/
 function decrypt(data, key) {
   var decipher = crypto.createDecipher('aes-256-cbc', key);
   //decipher.setAutoPadding(false);
   var decrypted_data = decipher.update(data,'hex','utf8');
-  decrypted_data += decipher.final('utf8');
-  return decrypted_data;
-}
-function decryptBuffer(data, key){
-  var decipher = crypto.createDecipher('aes-256-cbc',key);
-  var decrypted_data = Buffer.concat([decipher.update(data) , decipher.final()]);
+  try { decrypted_data += decipher.final('utf8');} 
+  catch(err){ decrypted_data = -1; }
   return decrypted_data;
 }
 
-function encryptBuffer(data, key) {
-  var cipher = crypto.createCipher('aes-256-cbc', key);
-  var encrypted_data = cipher.update(data,'utf8','hex');
-  encrypted_data += cipher.final('hex');
-  return encrypted_data;
-}
 function encrypt(data, key) {
   var cipher = crypto.createCipher('aes-256-cbc',key);
   var encrypted_data = Buffer.concat([cipher.update(data),cipher.final()]);
